pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinyseoai"
version = "0.2.0"
description = "TinySEO AI - Autonomous SEO Audit Agent with Multi-Agent AI"
authors = [
    { name = "Stalyn Disla" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["seo", "audit", "ai", "agent", "crawler", "analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.8",
    "platformdirs>=4.2",
    "httpx[http2]>=0.27",
    "trafilatura>=1.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "tldextract>=5.1",
    "yarl>=1.9",
    "robotexclusionrulesparser>=1.7",
    "pandas>=2.2",
    "numpy>=2.1",
    "jinja2>=3.1",
    "playwright>=1.48",
    "openpyxl>=3.1",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "mypy>=1.8",
    "ruff>=0.2",
    "black>=24.0",
    "isort>=5.13",
    "pre-commit>=3.6",
    "faker>=22.0",
]
build = ["pyinstaller>=6.10"]
ai = [
    "langchain>=0.1.0",
    "anthropic>=0.18.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",
]
all = ["tinyseoai[dev,build,ai]"]

[project.scripts]
tinyseoai = "tinyseoai.cli:app"

[project.urls]
Homepage = "https://tinyseoai.com"
Repository = "https://github.com/tinyseoai/tinyseoai-cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["tinyseoai*"]

# ==================== Black Configuration ====================
[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

# ==================== isort Configuration ====================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["tinyseoai"]

# ==================== Ruff Configuration ====================
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

ignore = [
    "E501",   # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # assert allowed in tests

# ==================== Bandit Configuration ====================
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]

# ==================== MyPy Configuration ====================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# ==================== Coverage Configuration ====================
[tool.coverage.run]
source = ["tinyseoai"]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
